cscl use seed file to format text fields (#2096)
* move some centerline cleaning to staging table

* macro/seed to apply text formatting rules

* update downstream columns and poc_loader

I was getting annoyed always needing double quotes in all my sql
queries for data qa. These fields don't make it into the outputs
anyways

* move lion dat formatting seed

* use text formatting seed/macro for sedat

+++ products/cscl/models/intermediate/segments/int__primary_segments.sql
@@ -38,12 +38,12 @@ primary_segments AS (
                 NULL AS boroughcode,
             {% endif -%}
             {% if source_layer == 'stg__shoreline' -%} 
-                NULL AS legacy_segmentid,
+                NULL::INT AS legacy_segmentid,
                 NULL AS from_level_code,
                 NULL AS to_level_code,
                 source.segment_seqnum,
             {% elif source_layer == 'stg__nonstreetfeatures' -%}
-                source.legacy_segmentid,
+                source.legacy_segmentid::INT,
                 NULL AS from_level_code,
                 NULL AS to_level_code,
                 seqnum.segment_seqnum,
